backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
ultralytics>=8.3.0
opencv-python-headless>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
mediapipe>=0.10.0
pydantic>=2.0.0
